+++ src/org/bds/test/unit/TestCasesFunctions.java
@@ -76,6 +76,53 @@ public class TestCasesFunctions extends TestCasesBase {
         // Function: Return different type of array (class)
         compileErrors(dir + "test70.bds", "Cannot cast A[] to B[]");
     }
+    @Test
+    public void test07() {
+        // Function definition
+        runAndCheck(dir + "run_07.bds", "j", 4L);
+    }
+
+    @Test
+    public void test08() {
+        // Function definition
+        runAndCheck(dir + "run_08.bds", "j", 5L);
+    }
+
+    @Test
+    public void test09() {
+        // Function with two return statements
+        runAndCheck(dir + "run_09.bds", "i", 5L);
+    }
+    @Test
+    public void test12() {
+        // Function: Variable inside a function shadowing a global variable
+        runAndCheck(dir + "run_12.bds", "i", 1L);
+    }
+
+    @Test
+    public void test13() {
+        // Function: Variable inside a function shadowing a global variable
+        runAndCheck(dir + "run_13.bds", "h", 8L);
+    }
+
+    @Test
+    public void test57() {
+        // Function: If with return
+        runAndCheck(dir + "run_57.bds", "z", 0L);
+    }
+
+    @Test
+    public void test58() {
+        // Function: recursion
+        runAndCheck(dir + "run_58.bds", "z", 0L);
+    }
+
+    @Test
+    public void test59() {
+        // Function: recursion
+        runAndCheck(dir + "run_59.bds", "z", -1L);
+    }
+
 
 
 }
