+++ src/org/bds/test/unit/TestCasesMap.java
@@ -27,5 +27,26 @@ public class TestCasesMap extends TestCasesBase {
         // Initialize with an empty map of objects
         compileOk(dir + "test72.bds");
     }
+    @Test
+    public void test52() {
+        // Hash assignment
+        runAndCheck(dir + "run_52.bds", "hash", "{ one => 1 }");
+    }
+    @Test
+    public void test55() {
+        // Hash: hasValue, size, hasKey
+        runAndCheck(dir + "run_55.bds", "hk1", "true");
+        runAndCheck(dir + "run_55.bds", "hk2", "false");
+        runAndCheck(dir + "run_55.bds", "hv1", "true");
+        runAndCheck(dir + "run_55.bds", "hv2", "false");
+        runAndCheck(dir + "run_55.bds", "hk3", "false");
+    }
+
+    @Test
+    public void test64() {
+        // Hash: Empty hash literal
+        runAndCheck(dir + "run_64.bds", "m", "{}");
+    }
+
 
 }
